+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # changelog
 
 * golang 1.21 is now the minimum version
+* :warning: `ReverseSlice` function in `basicalter` package is now deprecated  
+  use [slices.Reverse()](https://pkg.go.dev/slices#Reverse) from the standard [slices](https://pkg.go.dev/slices) library instead
 
 ## v0.11.0