+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * golang 1.21 is now the minimum version
 * :warning: `ReverseSlice` function in `basicalter` package is now deprecated  
   use [slices.Reverse()](https://pkg.go.dev/slices#Reverse) from the standard [slices](https://pkg.go.dev/slices) library instead
+* :warning: `EqualSlice` function in `basiccheck` package is now deprecated  
+  use [slices.Equal()](https://pkg.go.dev/slices#Equal) from the standard [slices](https://pkg.go.dev/slices) library instead
 
 ## v0.11.0