+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   use [slices.Equal()](https://pkg.go.dev/slices#Equal) from the standard [slices](https://pkg.go.dev/slices) library instead
 * :warning: `InSlice` function in `basiccheck` package is now deprecated  
   use [slices.Contains()](https://pkg.go.dev/slices#Contains) from the standard [slices](https://pkg.go.dev/slices) library instead
+* :warning: `OneInSliceWith` function in `basiccheck` package is now deprecated  
+  use [slices.ContainsFunc()](https://pkg.go.dev/slices#ContainsFunc) from the standard [slices](https://pkg.go.dev/slices) library instead
 
 ## v0.11.0